Fix watchdogThreshold and noiseLevel
The mask for getting and setting the noise floor was wrong.  There were actually
two masks.  Replaced them with a single mask, NOISE_FLOOR_MASK, that has 1 bits
where the noise floor field is within the register.

Setting the watchdogThreshold was using the wrong mask.  Inverted THRESH_MASK
so it has 1 bits where the watchdog field is within the register (you don't
need to invert it when using it in the code).